this is 1 part for java part 3
in this section the first function that creates the size of the canvas that is used in the web page by setting the height and the width of the canvas using the math function. next I used the math function again to generate a random number. and finally there is a function that uses a  random number