commit 2 for second part of part 3
in this section I had to call the ball class to create the ball on the screen using some math functions to set its place. then I had to use the draw function to make the ball and set its color from the random color generator used in the first part of these commits for part 3 one this was all set I used the update function to then move the ball around the screen.