section 2 of part 4 commit 4
in this section it is partially the same as part 1 what this does initially is it updates the size of the balls bouncing around on the screen. and also when the ball hits one of the walls it will change color and size. the new part that was added to this is the evil circle which will start at one point on the screen which is inputted into the code and its color is set to white  what else I did is also set key binds so that you are able to control the evil circle